+++ bundles/EcommerceFrameworkBundle/CartManager/CartFactory.php
@@ -45,18 +45,12 @@ class CartFactory implements CartFactoryInterface
 
         $resolver->setDefaults([
             'cart_class_name' => Cart::class,
-            'guest_cart_class_name' => SessionCart::class,
-            'cart_readonly_mode' => AbstractCart::CART_READ_ONLY_MODE_STRICT,
+            'guest_cart_class_name' => SessionCart::class
         ]);
 
         $resolver->setAllowedTypes('cart_class_name', 'string');
         $resolver->setAllowedTypes('guest_cart_class_name', 'string');
-        $resolver->setAllowedTypes('cart_readonly_mode', 'string');
 
-        $resolver->addAllowedValues('cart_readonly_mode', [
-            AbstractCart::CART_READ_ONLY_MODE_STRICT,
-            AbstractCart::CART_READ_ONLY_MODE_DEACTIVATED,
-        ]);
     }
 
     public function getCartClassName(EnvironmentInterface $environment): string
@@ -89,18 +83,6 @@ class CartFactory implements CartFactoryInterface
             $cart->setId($id);
         }
 
-        if ($cart instanceof AbstractCart) {
-            $cart->setCurrentReadonlyMode($this->options['cart_readonly_mode']);
-        }
-
         return $cart;
-    }
-
-    /**
-     * @return string
-     */
-    public function getCartReadOnlyMode(): string
-    {
-        return $this->options['cart_readonly_mode'];
     }
 }
