+++ bundles/EcommerceFrameworkBundle/CheckoutManager/CheckoutManagerFactory.php
@@ -20,6 +20,7 @@ namespace Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager;
 use Pimcore\Bundle\EcommerceFrameworkBundle\CartManager\CartInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\V7\CheckoutManager;
 use Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\V7\CheckoutManagerInterface;
+use Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\V7\HandlePendingPayments\CancelPaymentOrRecreateOrderStrategy;
 use Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\V7\HandlePendingPayments\ThrowExceptionStrategy;
 use Pimcore\Bundle\EcommerceFrameworkBundle\EnvironmentInterface;
 use Pimcore\Bundle\EcommerceFrameworkBundle\OrderManager\OrderManagerLocatorInterface;
@@ -149,7 +150,8 @@ class CheckoutManagerFactory implements CheckoutManagerFactoryInterface
         $resolver->setDefined('handle_pending_payments_strategy');
         $resolver->setAllowedTypes('handle_pending_payments_strategy', 'string');
 
-        $resolver->setDefault('handle_pending_payments_strategy', ThrowExceptionStrategy::class);
+        $resolver->setDefault('handle_pending_payments_strategy', CancelPaymentOrRecreateOrderStrategy::class);
+        $resolver->setRequired(['handle_pending_payments_strategy']);
     }
 
     public function createCheckoutManager(CartInterface $cart): CheckoutManagerInterface
@@ -176,10 +178,7 @@ class CheckoutManagerFactory implements CheckoutManagerFactoryInterface
             $this->eventDispatcher,
             $this->paymentProvider
         );
-
-        if ($checkoutManager instanceof \Pimcore\Bundle\EcommerceFrameworkBundle\CheckoutManager\V7\CheckoutManagerInterface) {
-            $checkoutManager->setHandlePendingPaymentsStrategy($this->handlePendingPaymentStrategy);
-        }
+        $checkoutManager->setHandlePendingPaymentsStrategy($this->handlePendingPaymentStrategy);
 
         $this->checkoutManagers[$cartId] = $checkoutManager;
